requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
authlib==1.5.2
beautifulsoup4==4.13.4
black==24.10.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coverage==7.8.0
cryptography==44.0.3
deprecated==1.2.18
dill==0.4.0
docstring-parser==0.16
fastapi==0.115.12
flake8==7.2.0
flake8-pyproject==1.2.3
frozenlist==1.6.0
gitdb==4.0.12
gitpython==3.1.44
google-adk==0.4.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-spanner==3.54.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mako==1.3.10
markupsafe==3.0.2
mccabe==0.7.0
mcp==1.7.1
multidict==6.4.4
mypy-extensions==1.1.0
narwhals==1.41.0
numpy==2.2.5
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
pyflakes==3.3.2
pyink==24.10.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
ruamel-yaml==0.18.10
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.8.0
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sqlalchemy-spanner==1.11.0
sqlparse==0.5.3
sse-starlette==2.3.4
starlette==0.46.2
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
toolbox-core==0.2.0
tornado==6.5.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
attrs==25.3.0
authlib==1.5.2
beautifulsoup4==4.13.4
black==24.10.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coverage==7.8.0
cryptography==44.0.3
deprecated==1.2.18
dill==0.4.0
docstring-parser==0.16
fastapi==0.115.12
flake8==7.2.0
flake8-pyproject==1.2.3
frozenlist==1.6.0
gitdb==4.0.12
gitpython==3.1.44
google-adk==0.4.0
google-api-core==2.25.1
google-api-python-client==2.169.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-spanner==3.54.0
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mako==1.3.10
markupsafe==3.0.2
mccabe==0.7.0
mcp==1.7.1
multidict==6.4.4
mypy-extensions==1.1.0
narwhals==1.41.0
numpy==2.2.5
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
pyflakes==3.3.2
pyink==24.10.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.24.0
rsa==4.9.1
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.8.0
shapely==2.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.40
sqlalchemy-spanner==1.11.0
sqlparse==0.5.3
sse-starlette==2.3.4
starlette==0.46.2
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.2
toolbox-core==0.2.0
tornado==6.5.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.23.0
